extract TaskList from TaskDisplay
