+++ src/pages/Browse/TaskDisplay/TaskDisplay.jsx
@@ -1,55 +1,10 @@
 import React, { useEffect } from 'react';
-import classNames from 'classnames/bind';
-import { Redirect, Route, Switch, useHistory, useRouteMatch } from 'react-router-dom';
+import { Redirect, Route, Switch, useRouteMatch } from 'react-router-dom';
 
 import styles from './TaskDisplay.module.scss';
 
 import TaskDetail from '../TaskDetail';
-import Avatar from '../../../components/Avatar';
-
-const cx = classNames.bind(styles);
-
-function TaskListItem({
-  onClick,
-  task: {
-    title, status, budget, postedBy, location, due,
-  },
-}) {
-  return (
-    <div
-      className={cx(
-        'task_list_item',
-        status.toLowerCase(),
-      )}
-      onClick = {onClick}
-    >
-      <div className={styles.header} >
-        <div className={styles.title} >
-          {title}
-        </div>
-        <div className={styles.budget} >
-          ${budget}
-        </div>
-      </div>
-      <div className={styles.content} >
-        <div className={styles.avatar_container} >
-          <Avatar avatarUrl={postedBy.avatar || null} />
-        </div>
-        <div className={styles.location} >
-          {location}
-        </div>
-        <div className={styles.due} >
-          {due}
-        </div>
-      </div>
-      <div className={styles.footer} >
-        <div className={styles.status} >
-          {status}
-        </div>
-      </div>
-    </div>
-  );
-}
+import TaskList from '../TaskList';
 
 export default function TaskDisplay({
   taskList,
@@ -57,7 +12,6 @@ export default function TaskDisplay({
   lastScroll,
   saveScroll,
 }) {
-  const history = useHistory();
   const { path } = useRouteMatch();
 
   useEffect(() => {
@@ -82,23 +36,14 @@ export default function TaskDisplay({
   const defaultTask = taskList.length && taskList[0].id;
 
   return (
-    <div className={styles.task_list_wrapper}>
-      <div
-        id="task-list"
-        className = {styles.task_list}
-        onScroll={handleScroll}
-      >
-        {taskList.map((task) => (
-          <TaskListItem
-            key={task.id}
-            task={task}
-            onClick={() => history.push(`${path}/${task.id}`)}
-          />
-        ))}
-      </div>
+    <div className={styles.container}>
+      <TaskList
+        taskList={taskList}
+        handleScroll={handleScroll}
+      />
       <Switch>
         <Route path={`${path}/:taskId`} >
-          <div className={styles.task_detail_wrapper} >
+          <div className={styles.right} >
             <TaskDetail taskList={taskList} />
           </div>
         </Route>
