+++ src/pages/Browse/TaskDisplay/TaskDisplay.module.scss
@@ -1,9 +1,4 @@
-@import '../../../components/scss-partial/status-color-code';
-
-$text-color: #292b32;
-$content-color: #545a77;
-
-.task_list_wrapper{
+.container{
   position: relative;
   display: flex;
   justify-content: space-between;
@@ -15,136 +10,6 @@ $content-color: #545a77;
   word-wrap: break-word;
 }
 
-.task_list{
-  background-color: lighten(grey, 40%);
-  display: flex;
-  flex-flow: column;
-  overflow-y: scroll;
-  width: 300px;
-  max-height: 100%;
-}
-
-.task_detail_wrapper{
+.right{
   flex: 1 1;
-}
-
-.task_list_item{
-  position: relative;
-  background-color: white;
-  display: flex;
-  flex-direction: column;
-  margin-top: 10px;
-  margin-right: 7px;
-  padding: 0 12px;
-  text-align: left;
-  border-radius: 4px;
-  cursor: pointer;
-
-  &::before{
-    content: '';
-    position: absolute;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    width: 4px;
-    display: block;
-    border-top-left-radius: 4px;
-    border-bottom-left-radius: 4px;
-  }
-
-  &:hover{
-    box-shadow: 3px 5px 8px rgba(187,194,220,.6);
-  }
-
-  .header{
-    display: flex;
-    position: relative;
-    margin-top: 10px;
-    margin-bottom: 5px;
-
-    .title{
-      font-weight: 300;
-      flex-grow: 1;
-      font-size: 17px;
-      line-height: 20px;
-      color: $text-color;
-      word-break: break-word;
-    }
-
-    .budget{
-      font-weight: 500;
-      float: right;
-      font-size: 24px;
-      margin-left: 10px;
-      line-height: 24px;
-      color: $text-color;
-      float: right;
-    }
-  }
-
-  .content{
-    display: flex;
-    flex-direction: column;
-    position: relative;
-    margin-top: 5px;
-    margin-bottom: 10px;
-    color: $content-color;
-    min-height: 34px;
-
-    .location, .due{
-      margin-left: 5px;
-      display: inline-block;
-      vertical-align: middle;
-      font-size: 12px;
-    }
-
-    .avatar_container{
-      position: absolute;
-      right: 0;
-      width: 30px;
-      height: 30px;
-    }
-  }
-
-  .footer{
-    height: 32px;
-    line-height: 32px;
-    border-top: 1px solid #e7ebfb;
-
-    .status{
-      text-transform: uppercase;
-      font-weight: 700;
-      font-size: 11px;
-    }
-  }
-}
-
-@mixin status-color-code($status-color){
-  &:hover{
-    .footer{
-      .status{
-        color: $hover-color;
-      }
-    }
-  }
-  &::before{
-    background-color: $status-color;
-  }
-  .footer{
-    .status{
-      color: $status-color;
-    }
-  }
-}
-
-.open{
-  @include status-color-code($open-color);
-}
-
-.assigned{
-  @include status-color-code($assigned-color);
-}
-
-.completed{
-  @include status-color-code($completed-color);
 }
