+++ src/pages/Browse/TaskList/index.js
@@ -1,0 +1,1 @@
+export { default } from './TaskList';
